requirements.txt
pytest==7.3.1
requests==2.31.0
python-dotenv==1.0.0